refactor(sqlite): restructure functions into submodules by category
Split the monolithic functions.py into organized submodules following the
PostgreSQL backend pattern. This improves code organization and maintainability.

Changes:
- Create functions/ subdirectory with category-based submodules:
  - functions/string.py: String manipulation functions (substr, instr, etc.)
  - functions/datetime.py: Date/time functions (date, time, datetime, etc.)
  - functions/math.py: Mathematical functions (random, abs, sign, total)
  - functions/blob.py: BLOB functions (zeroblob, randomblob)
  - functions/system.py: System functions (typeof, quote, last_insert_rowid, changes)
  - functions/conditional.py: Conditional functions (iif)
- Create functions/__init__.py that aggregates all function exports
- Delete original functions.py
- Update __init__.py to import from functions subpackage

This structure allows importing individual functions:
  from rhosocial.activerecord.backend.impl.sqlite.functions import substr
  from rhosocial.activerecord.backend.impl.sqlite.functions.datetime import date_func

Or importing from the parent package (backward compatible):
  from rhosocial.activerecord.backend.impl.sqlite import substr
