+++ src/rhosocial/activerecord/backend/impl/sqlite/__init__.py
@@ -76,26 +76,26 @@ from .functions import (
     unicode,
     hex,
     unhex,
-    zeroblob,
-    randomblob,
     soundex,
     group_concat,
-    total,
+    trim_sqlite,
+    ltrim,
+    rtrim,
     date_func,
     time_func,
     datetime_func,
     julianday,
     strftime_func,
-    typeof,
-    quote,
     random_func,
     abs_sql,
     sign,
+    total,
+    zeroblob,
+    randomblob,
+    typeof,
+    quote,
     last_insert_rowid,
     changes,
-    trim_sqlite,
-    ltrim,
-    rtrim,
     iif,
 )
